chunking/requirements.txt
smart-chunker
accelerate
python-magic